+++ provider/pkg/gen/dotnet-templates/helm/v3/Chart.cs
@@ -397,9 +397,9 @@ namespace Pulumi.Kubernetes.Helm.V3
             public string? Repo { get; set; }
             [JsonPropertyName("chart")]
             public string? Chart { get; set; }
-            [JsonPropertyName("version")]
-            public string? KubeVersion { get; set; }
             [JsonPropertyName("kube_version")]
+            public string? KubeVersion { get; set; }
+            [JsonPropertyName("version")]
             public string? Version { get; set; }
             [JsonPropertyName("fetch_opts")]
             public JsonOptsFetch? FetchOptions { get; set; }